mypy: use stricter settings (#52)
* mypy: use stricter settings

---------

+++ qiskit_aqt_provider/aqt_job.py
@@ -466,7 +466,7 @@ def _shot_to_int(
     else:
         creg = fluorescence_states.copy()
 
-    return (np.left_shift(1, np.arange(len(creg))) * creg).sum()
+    return int((np.left_shift(1, np.arange(len(creg))) * creg).sum())
 
 
 def _format_counts(
